move types and constants in separate files
