+++ indexers/gemini-3h/staking/src/mappings/constants.ts
@@ -1,0 +1,1 @@
+export const SHARES_CALCULATION_MULTIPLIER = BigInt(1000000000000);
